GEODE-10474: Update logomark after ASF rebrand
GEODE-10474: Project website does not reflect updated branding

The ASF recently launched a new brand and visual identity, which incudes
a new logomark. This commit replaces the former logomark on the Geode
project home page with the new one in order to comply with ASF brand
guidelines. It also updates the copyright year to 2025.

Signed-off-by: Bryan Behrenshausen <bryan.behrenshausen@sas.com>